requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.2
django-environ==0.4.5
djangorestframework-simplejwt==4.6.0
Markdown==3.3.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
django-heroku